pyproject.toml
[project]
name = "painter"
version = "0.1.0"
description = "Agent that learns to paint by iterating on a browser canvas."
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "pillow>=10",
    "scikit-image>=0.22",
    "scipy>=1.11",
    "numpy>=1.26",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
